Part 4B – Sorting the collection
Created a new VisitorComparator class with a method that sorts visitors by name and age.
Updated the sortRideHistory() method in the Ride class to use this function.
Updated some outputs of the other Ride class methods.
Updated partFourB() of the AssignmentTwo class to instantiate some visitors, add them to the ride history, print and sort the ride history.
Made all the methods in Assignment Two static methods.