Part 7 – Reading from a file
Updated the exportRideHistory() method to write a different output so the file can be read more easily.
Added the importRideHistory() method to the Ride class that imports visitors from the file as objects and adds them back to the RideHistory LL.
Updated partSeven() in the AssignmentTwo class to import a ride's history, and then output the number of visitors stored in the history and print the history.