Updates and improvements
Updated setter methods in Employee class setEmploymentType(), and setRole(), and setTicketType() in Visitor class.
Updated Ride class to utilise constructor overloading, where the maxRider parameter is not given.
Fixed issue in getNumOfCycles() to stop when the queue is empty.
Updated the numberOfVisitors() method to return the number of visitors as an integer.
Added detailed / improved comments and output of the program.